Moved top-level windows out of scripts and into dbsgui sub-package
