+++ neuroport_dbs/dbsgui/process.py
@@ -1,9 +1,7 @@
-import sys
-import argparse
 from pathlib import Path
 from qtpy import QtCore, QtWidgets, QtGui
-from neuroport_dbs.settings import defaults, parse_ini_try_numeric, locate_ini
-from neuroport_dbs.SettingsDialog import SettingsDialog
+from neuroport_dbs.settings import defaults, locate_ini
+from neuroport_dbs.dbsgui.widgets.SettingsDialog import SettingsDialog
 from serf.tools.db_wrap import DBWrapper, ProcessWrapper
 
 
@@ -234,23 +232,3 @@ class ProcessGUI(QtWidgets.QMainWindow):
                                   f"background-color : {feat_facecolor}; "
                                   f"border-color : {feat_facecolor}; "
                                   "border-width: 2px}")
-
-
-def main(**kwargs):
-    app = QtWidgets.QApplication(sys.argv)
-    window = ProcessGUI(**kwargs)
-    window.show()
-    timer = QtCore.QTimer()
-    timer.timeout.connect(window.update)
-    timer.start(100)
-
-    if (sys.flags.interactive != 1) or not hasattr(QtCore, 'PYQT_VERSION'):
-        sys.exit(app.exec_())
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(prog="ProcessGUI",
-                                     description="Manage subprocesses to store segments and compute features.")
-    parser.add_argument('-i', '--ini_file', nargs='?', help="Path to ini settings file.")
-    args = parser.parse_args()
-    main(**args.__dict__)
